requirements.txt
Django==1.6
MarkupSafe==0.15
South==0.8.2
bleach==1.2.2
coverage==3.7.1
crispy-forms-foundation==0.2.3
django-crispy-forms==1.3.2
django-pipeline==1.3.15
pep8==1.4.6
pyflakes==0.7.3
pillow
pytz
gitpython
factory-boy==2.2.1
django-discover-runner==0.4
validate_email
pyDNS